LayoutTests/imported/w3c/web-platform-tests/html/user-activation/propagation-crossorigin.sub.html

<!DOCTYPE html>
<html>
  <head>
    <meta name="timeout" content="long" />
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="/resources/testdriver.js"></script>
    <script src="/resources/testdriver-vendor.js"></script>
    <script src="resources/utils.js"></script>
  </head>
  <body></body>
  <script>
    async function requestUserActivationSnapshot(iframe, type = "report") {
      const msgId = `${iframe.id}-${type}`;
      const msg = `{"type": "${msgId}"}`;
      iframe.contentWindow.postMessage(msg, "*");
      const data = await receiveMessage(msgId);
      return data;
    }

    test(() => {
      assert_false(navigator.userActivation.isActive, "top isActive");
      assert_false(navigator.userActivation.hasBeenActive, "top hasBeenActive");
    }, "Initial user activation states");

    promise_test(async () => {
      const child2LoadedPromise = receiveMessage("child-two-loaded");
      const child1 = await attachIframe(
        "http://{{hosts[alt][]}}:{{ports[http][0]}}/html/user-activation/resources/child-one.html"
      );
      await receiveMessage("child-one-loaded");
      child1.id = "child-one";

      const childXo = await attachIframe(
        "http://{{hosts[alt][]}}:{{ports[http][1]}}/html/user-activation/resources/propagation-crossorigin-child.sub.html"
      );
      childXo.id = "child-crossorigin";
      await receiveMessage("child-crossorigin-loaded");

      // before click
      for (const frame of [child1, childXo]) {
        const { isActive, hasBeenActive } = await requestUserActivationSnapshot(
          frame
        );
        assert_false(isActive, `Frame ${frame.src} isActive`);
        assert_false(hasBeenActive, `Frame ${frame.src} hasBeenActive`);
      }

      // Ensure the nestedIframe is fully loaded before we proceed
      await child2LoadedPromise;

      // Check the childXo's sub frame
      childXo.contentWindow.postMessage(
        '{"type": "child-crossorigin-iframe-report"}',
        "*"
      );
      const nestedFrameReport = await receiveMessage("child-two-report");
      assert_false(nestedFrameReport.isActive, `nested iframe isActive`);
      assert_false(
        nestedFrameReport.hasBeenActive,
        `nested iframe hasBeenActive`
      );

      // Let's click!
      await test_driver.click(childXo);

      // let's gather the reports for the new user activation states
      const reports = {
        child1: await requestUserActivationSnapshot(child1),
        childXo: await requestUserActivationSnapshot(childXo),
      };
      childXo.contentWindow.postMessage(
        '{"type": "child-crossorigin-iframe-report"}',
        "*"
      );
      reports.nestedFrame = await receiveMessage("child-two-report");

      // Check child 1 (not clicked!)
      assert_false(reports.child1.isActive, `child1 isActive`);
      assert_false(reports.child1.hasBeenActive, `child1 hasBeenActive`);

      // check childXo, was clicked
      assert_true(reports.childXo.isActive, `childXo isActive`);
      assert_true(reports.childXo.hasBeenActive, `childXo hasBeenActive`);

      // Finally, check the nested cross-origin iframe (not clicked) but is same origin
      assert_true(reports.nestedFrame.isActive, `nestedFrame isActive`);
      assert_true(
        reports.nestedFrame.hasBeenActive,
        `nestedFrame hasBeenActive`
      );
    });

    test(() => {
      assert_false(navigator.userActivation.isActive, "top isActive");
      assert_false(navigator.userActivation.hasBeenActive, "top hasBeenActive");
    }, "Final user activation states");
  </script>
</html>
